:recycle: [Refactor]: bookmark를 features 레이어로 분리
