+++ apps/client/src/features/bookmark/index.ts
@@ -1,0 +1,2 @@
+export { Bookmark } from './ui/Bookmark';
+export type { BookmarkData } from './model/types';
