+++ apps/client/src/features/bookmark/model/types.ts
